+++ NEWS.md
@@ -3,6 +3,7 @@
 ## v0.7.1
 
 - Remove unrar.cffi support.
+- Test for unrar executable for clearer errors.
 
 ## v0.7.0